Add dummy implementations of all entity mixins
