+++ betty/tests/ancestry/test_has_notes.py
@@ -4,27 +4,16 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from betty.ancestry.has_notes import HasNotes
 from betty.ancestry.note import Note
-from betty.model import EntityDefinition
+from betty.test_utils.ancestry.has_notes import DummyHasNotes
 from betty.test_utils.json.linked_data import assert_dumps_linked_data
 from betty.test_utils.locale.localizable import (
-    DUMMY_COUNTABLE_LOCALIZABLE,
     DUMMY_LOCALIZABLE,
 )
 
 if TYPE_CHECKING:
+    from betty.ancestry.has_notes import HasNotes
     from betty.serde.dump import Dump, DumpMapping
-
-
-@EntityDefinition(
-    "dummy-has-notes",
-    label=DUMMY_LOCALIZABLE,
-    label_plural=DUMMY_LOCALIZABLE,
-    label_countable=DUMMY_COUNTABLE_LOCALIZABLE,
-)
-class DummyHasNotes(HasNotes):
-    pass
 
 
 class TestHasNotes:
