+++ betty/tests/project/load/test_jobs.py
@@ -1,17 +1,13 @@
 import pytest
 from aioresponses import aioresponses
 
-from betty.ancestry.has_links import HasLinks
 from betty.ancestry.link import Link
 from betty.app import App
 from betty.locale.localize import DEFAULT_LOCALIZER
 from betty.project import Project, ProjectContext
 from betty.project.load.jobs import PopulateLink
+from betty.test_utils.ancestry.has_links import DummyHasLinks
 from betty.test_utils.job import do
-
-
-class DummyHasLinks(HasLinks):
-    pass
 
 
 class TestPopulateLink:
